code refactoring & adding Dockerfile & adding Makefile
