+++ go.mod
@@ -1,3 +1,6 @@
 module github.com/ansavin/system-monitoring
 
 go 1.16
+
+require os/linux v1.0.0
+replace os/linux => ./os/linux
