+++ os/linux/go.mod
@@ -1,0 +1,3 @@
+module linux
+
+go 1.16
